refactor(optimizer): split visual-state seed cluster into RecordingOptimizer.SeedEvents.cs
